requirements.txt
requests
pandas
gspread
google-auth
google-generativeai
python-dotenv
